change motor_number to be an uint8_t from an int

+++ src/product/motor.hpp
@@ -16,7 +16,7 @@ private:
   AF_DCMotor *_motor_ptr;
   uint8_t _speed = 0;
 public:
-  motor(uint8_t, int);
+  motor(uint8_t, uint8_t);
   ~motor();
   void run_forward(bool ramp_up = false);
   void run_backward(bool ramp_up = false);
